split gitlab project ref inflight tracking
